Refreshing contact list functionality added.
With this update:

If the user is using the app for the first time, after entering his
phone number and name, users' all contacts are sent to the server, and
checked if the contacts are using the app. And contacts page updated
with the received info(onApp) from the server. Also this data is stored
in the phone's database.

If the user is not using the app for the first time, after connecting to
the server with stored info(phone number and name), contacts(with onApp
fields) are loaded from the phone's database and displayed in Contacts
page.

-----

If user clicks on the "Refresh list" button on Contacts page, all
contacts are again sent to the server, checked, old contacts(with onApp
fields) on the phone's database replaced with the new data, and new data displayed in the contacts page.