Social share support added, working tree cleaned. Changes: - Social share support added.
If you click on the "Share the app" button on the HomePage, it takes you
to the share page. On the share page, you can click on images, and a
customizable message can be shared on Facebook, Twitter, Whatsapp or
with e-mail. (Check SocialSharing Plugin info page for additional social
media options.)

- Working tree cleaned.(Some unused files are deleted.) Some images
	added for social sharing.
- README.md updated according to the latest function definitons, and
	some used library informations added.